+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentDisplayer/index.jsx
@@ -1,16 +1,25 @@
-import { Button, Space } from 'antd';
-import React, { PureComponent } from 'react';
+import { Button } from 'antd';
+import React from 'react';
 import ReactToPrint from 'react-to-print';
 
-import { isArray, whetherString } from 'easy-soft-utility';
+import { isArray, isFunction, toMd5, whetherString } from 'easy-soft-utility';
 
-import { CenterBox, iconBuilder } from 'antd-management-fast-component';
+import {
+  BaseComponent,
+  buildButton,
+  CenterBox,
+  EverySpace,
+  iconBuilder,
+  PageExtra,
+} from 'antd-management-fast-component';
 
 import { DocumentContent } from './DocumentContent';
 
+const { ToolBar } = PageExtra;
+
 const colorDefault = '#000';
 
-class DocumentDisplayer extends PureComponent {
+class DocumentDisplayer extends BaseComponent {
   componentRef = null;
 
   constructor(properties) {
@@ -19,15 +28,32 @@ class DocumentDisplayer extends PureComponent {
     this.state = {
       ...this.state,
       designMode: true,
+      schemaTag: '',
+      schemaStorage: [],
     };
+  }
+
+  // eslint-disable-next-line no-unused-vars
+  static getDerivedStateFromProps(nextProperties, previousState) {
+    const { schema } = nextProperties;
+    const { schemaTag } = previousState;
+
+    if (toMd5(JSON.stringify(schema || [])) != schemaTag) {
+      return {
+        schemaTag: toMd5(JSON.stringify(schema || [])),
+        schemaStorage: schema || [],
+      };
+    }
+
+    return null;
   }
 
   setComponentRef = (reference) => {
     this.componentRef = reference;
   };
 
-  reset = () => {
-    const { schema, onChange: onChangeCallback } = this.props;
+  initializeDesign = () => {
+    const { schema } = this.props;
 
     if (isArray(schema)) {
       const schemaAdjust = schema.map((o) => {
@@ -49,9 +75,29 @@ class DocumentDisplayer extends PureComponent {
         };
       });
 
-      onChangeCallback(schemaAdjust);
+      this.setState({ schemaStorage: schemaAdjust });
     } else {
-      onChangeCallback(schema);
+      this.setState({ schemaStorage: [] });
+    }
+  };
+
+  reset = () => {
+    const { schema } = this.props;
+
+    this.setState({ schemaStorage: schema || [] });
+  };
+
+  onChange = (data) => {
+    this.setState({ schemaStorage: data || [] });
+  };
+
+  save = () => {
+    const { onSave } = this.props;
+
+    if (isFunction(onSave)) {
+      const { schemaStorage } = this.state;
+
+      onSave(schemaStorage || []);
     }
   };
 
@@ -89,7 +135,6 @@ class DocumentDisplayer extends PureComponent {
   render() {
     const {
       values,
-      schema,
       style,
       color,
       labelColumnWidth,
@@ -100,13 +145,12 @@ class DocumentDisplayer extends PureComponent {
       title,
       titleContainerStyle,
       titleStyle,
-      onChange: onChangeCallback,
     } = this.props;
-    const { designMode } = this.state;
+    const { designMode, schemaStorage } = this.state;
 
     const p = {
       values,
-      schema,
+      schema: schemaStorage,
       style,
       color,
       labelColumnWidth,
@@ -117,72 +161,106 @@ class DocumentDisplayer extends PureComponent {
       title,
       titleContainerStyle,
       titleStyle,
-      onChange: onChangeCallback,
+      onChange: this.onChange,
     };
 
     return (
       <>
-        <div
-          style={{
-            position: 'relative',
-            height: '100%',
-          }}
-        >
-          <div style={{ position: 'absolute', top: '00px', right: '80px' }}>
-            <Space>
-              {designMode ? (
-                <Button
-                  icon={iconBuilder.form()}
-                  size="small"
-                  onClick={this.closeDesignMode}
-                >
-                  关闭设计
-                </Button>
-              ) : null}
+        <div>
+          <CenterBox>
+            <div>
+              <EverySpace size={10} direction="horizontal" />
 
-              {designMode ? (
-                <Button
-                  icon={iconBuilder.redo()}
-                  size="small"
-                  onClick={this.reset}
-                >
-                  重置设置
-                </Button>
-              ) : null}
-
-              {designMode ? null : (
-                <Button
-                  icon={iconBuilder.form()}
-                  size="small"
-                  onClick={this.openDesignMode}
-                >
-                  开启设计
-                </Button>
-              )}
-
-              {designMode ? null : (
-                <ReactToPrint
-                  content={this.renderPrintContent}
-                  documentTitle="文档"
-                  // onAfterPrint={this.handleAfterPrint}
-                  // onBeforeGetContent={this.handleOnBeforeGetContent}
-                  // onBeforePrint={this.handleBeforePrint}
-                  removeAfterPrint
-                  trigger={this.renderTrigger}
+              <div style={{ borderRadius: '8px', overflow: 'hidden' }}>
+                <ToolBar
+                  stick={false}
+                  title="操作栏"
+                  tools={[
+                    {
+                      hidden: !designMode,
+                      component: buildButton({
+                        text: '关闭设计模式',
+                        size: 'small',
+                        icon: iconBuilder.closeCircle(),
+                        handleClick: () => {
+                          this.closeDesignMode();
+                        },
+                      }),
+                    },
+                    {
+                      hidden: !designMode,
+                      component: buildButton({
+                        text: '初始化设计',
+                        size: 'small',
+                        icon: iconBuilder.clear(),
+                        handleClick: () => {
+                          this.initializeDesign();
+                        },
+                      }),
+                    },
+                    {
+                      hidden: !designMode,
+                      component: buildButton({
+                        text: '重置设计',
+                        size: 'small',
+                        icon: iconBuilder.redo(),
+                        handleClick: () => {
+                          this.reset();
+                        },
+                      }),
+                    },
+                    {
+                      hidden: !designMode,
+                      component: buildButton({
+                        text: '保存设计',
+                        size: 'small',
+                        type: 'primary',
+                        icon: iconBuilder.save(),
+                        handleClick: () => {
+                          this.save();
+                        },
+                      }),
+                    },
+                    {
+                      hidden: designMode,
+                      component: buildButton({
+                        text: '开启设计模式',
+                        size: 'small',
+                        icon: iconBuilder.desktop(),
+                        handleClick: () => {
+                          this.openDesignMode();
+                        },
+                      }),
+                    },
+                    {
+                      hidden: designMode,
+                      component: (
+                        <ReactToPrint
+                          content={this.renderPrintContent}
+                          documentTitle="文档"
+                          // onAfterPrint={this.handleAfterPrint}
+                          // onBeforeGetContent={this.handleOnBeforeGetContent}
+                          // onBeforePrint={this.handleBeforePrint}
+                          removeAfterPrint
+                          trigger={this.renderTrigger}
+                        />
+                      ),
+                    },
+                  ]}
                 />
-              )}
-            </Space>
-          </div>
+              </div>
 
-          <div>
-            <CenterBox>
+              <EverySpace size={10} direction="horizontal" />
+
               <DocumentContent
                 ref={this.setComponentRef}
                 {...p}
                 designMode={designMode}
               />
-            </CenterBox>
-          </div>
+            </div>
+          </CenterBox>
+
+          <EverySpace size={10} direction="horizontal" />
         </div>
       </>
     );