+++ packages/example-antd-management-fast-framework/src/pages/Interaction/DrawerView/index.jsx
@@ -423,6 +423,7 @@ class DrawerView extends BaseView {
         <SimpleMultiPageDrawer />
 
         <SimpleVerticalFlexDrawer
+          maskClosable
           afterClose={() => {
             logDebug({}, 'trigger afterClose');
           }}